Move InjectableReferenceTest to full + move back to using decorators
This test naturally belongs to the `full` group, because it verifies
behavior of `BeanManager.getInjectableReference()`, which is only
required to work in CDI Full. The `BeanContainer` interface does not
define this method.

Furthermore, this test was previously modified to use an interceptor
instead of a decorator. This commit restores the usage of a decorator
to avoid confusing method name and specification reference.
