+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/BasicLogger.java
@@ -9,13 +9,26 @@
  * You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,  
+ * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint;
 
-public interface Logger {
-    String log(String message);
+import jakarta.enterprise.context.Dependent;
+
+@Dependent
+public class BasicLogger implements Logger {
+
+    private static String message;
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void log(String message) {
+        BasicLogger.message = message;
+    }
+
 }
