+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/Logger.java
@@ -9,28 +9,13 @@
  * You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,  
+ * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint;
 
-import jakarta.enterprise.context.Dependent;
-
-@Dependent
-@SomeBinding
-public class BasicLogger implements Logger {
-
-    private static String message;
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String log(String message) {
-        BasicLogger.message = message;
-        return message;
-    }
-
+public interface Logger {
+    void log(String message);
 }
