--- impl/src/main/java/org/jboss/cdi/tck/tests/lookup/injectionpoint/SomeBinding.java
@@ -1,16 +1,0 @@
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
-
-import jakarta.interceptor.InterceptorBinding;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Inherited
-@InterceptorBinding
-@Target({ ElementType.TYPE, ElementType.METHOD, ElementType.CONSTRUCTOR })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SomeBinding {
-}
